LEP-508 Docker images tagging - staging_mtr and production
staging_mtr and production need to have their own tags, so that they do not share an ECR lifecycle rule,
otherwise lots of builds of one will cause the lifecycle rule to delete the in-use image for the other.

Also rename 'dev' to 'uat' so that we consistently tag by environment name. Will change the lifecycle rules to match.