Update readme to clarify integration test setup
Remove FUTURE_THRESHOLD_FILE from .env setup, as this isn't required for the spreadsheet integration tests to pass,
And will cause tests to fail when running bundle exec rspec or bundle exec cucumber.

This could be confusing for new developers.

Move the instructios for using the threshold file when running locally to the Test threshold data section instead.