EL-2156: Convert passported (DDD) tests to cucumber (#759)
* EL-2156: Convert passported (DDD) spreadsheet tests to cucumber

If applied, this converts passported (DDD) spreadsheet tests to cucumber.
- Added a new cucumber step definition "I add a benefits regular_transactions of {float} per month of credit"
- This also formats the BBB & CCC cucumber tests

* Disabling `undercover`

In addition this change disables undercover code coverage as it was behaving oddly when adding a new method to features/step_definitions/api_request_steps.rb. The coverage complaint realted to an existing method already used in a test, not the new one added. Before disabling undercover I had tried the following:

- adding a filter to spec_helper to ignore step definitions & features folder
- adding # :nocov: to various parts of features/step_definitions/api_request_steps.rb